+++ tnts_forge/src/main/java/com/tnts/ModCreativeTabs.java
@@ -25,6 +25,8 @@ public class ModCreativeTabs {
                         output.accept(new ItemStack(ModItems.TNT_ARROW.get()));
                         output.accept(new ItemStack(ModItems.GRENADE.get()));
                         output.accept(new ItemStack(ModItems.TNT_CHESTPLATE.get()));
+                        output.accept(new ItemStack(ModItems.TNT_BOOTS.get()));
+                        output.accept(new ItemStack(ModItems.TNT_HELMET.get()));
                         output.accept(new ItemStack(ModItems.TNT_PICKAXE.get()));
                         output.accept(new ItemStack(ModItems.TNT_MANUAL.get()));
                         output.accept(new ItemStack(ModItems.TNT_DISC.get()));